+++ src/main/java/persistence/entity/loader/OneToManyEntityLoader.java
@@ -1,0 +1,42 @@
+package persistence.entity.loader;
+
+import java.util.List;
+import jdbc.JdbcTemplate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import persistence.entity.EntityLoader;
+import persistence.sql.QueryGenerator;
+
+public class OneToManyEntityLoader implements EntityLoader {
+    private final Logger log = LoggerFactory.getLogger(OneToManyEntityLoader.class);
+    private final JdbcTemplate jdbcTemplate;
+    private final QueryGenerator queryGenerator;
+    private final EntityMapper oneToManyEntityMapper;
+
+    public OneToManyEntityLoader(JdbcTemplate jdbcTemplate, QueryGenerator queryGenerator,
+                                 EntityMapper oneToManyEntityMapper) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.queryGenerator = queryGenerator;
+        this.oneToManyEntityMapper = oneToManyEntityMapper;
+    }
+    public <T> T find(Class<T> tClass, Object id) {
+        final String query = queryGenerator
+                .select()
+                .findByIdQuery(id);
+
+        log.info(query);
+
+        return jdbcTemplate.queryForObject(query,
+                (resultSet) -> oneToManyEntityMapper.findMapper(tClass, resultSet));
+    }
+
+    public <T> List<T> findAll(Class<T> tClass) {
+        final String query = queryGenerator.select().findAllQuery();
+
+        log.info(query);
+
+        return jdbcTemplate.queryForAll(query,
+                (resultSet) -> oneToManyEntityMapper.findAllMapper(tClass, resultSet));
+    }
+
+}
