+++ src/main/java/persistence/entity/loader/OneToManyLazyLoader.java
@@ -1,0 +1,63 @@
+package persistence.entity.loader;
+
+import java.util.List;
+import jdbc.JdbcTemplate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import persistence.entity.EntityLoader;
+import persistence.sql.QueryGenerator;
+
+public class OneToManyLazyLoader implements EntityLoader {
+    private final Logger log = LoggerFactory.getLogger(OneToManyLazyLoader.class);
+    private final JdbcTemplate jdbcTemplate;
+    private final QueryGenerator queryGenerator;
+    private final EntityMapper entityMapper;
+
+    public OneToManyLazyLoader(JdbcTemplate jdbcTemplate, QueryGenerator queryGenerator,
+                               EntityMapper entityMapper) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.queryGenerator = queryGenerator;
+        this.entityMapper = entityMapper;
+    }
+
+    public <T> T find(Class<T> tClass, Object id) {
+        final String query = queryGenerator
+                .select()
+                .findByIdQuery(id);
+
+        log.info(query);
+
+        return jdbcTemplate.queryForObject(query,
+                (resultSet) -> entityMapper.findMapper(tClass, resultSet));
+    }
+
+    public <T> List<T> findAll(Class<T> tClass) {
+        final String query = queryGenerator.select().findAllQuery();
+
+        log.info(query);
+
+        return jdbcTemplate.query(query,
+                (resultSet) -> entityMapper.findMapper(tClass, resultSet));
+    }
+
+
+    //    public <T> T resultSetLazyManyEntity(Class<T> tClass, ResultSet resultSet, EntityLoader entityLoader, Callback callback) {
+//        final T instance = getInstance(tClass);
+//        final EntityMeta entityMeta = EntityMeta.from(tClass);
+//
+//        for (EntityColumn entityColumn : entityMeta.getEntityColumns()) {
+//            final String columnNameSignature = entityColumn.getColumnNameSignature(entityMeta.getTableName(), 0);
+//            final Object resultSetColumn = getLoadValue(resultSet, entityColumn, columnNameSignature);
+//            setFieldValue(instance, entityColumn.getFieldName(), resultSetColumn);
+//        }
+//
+//        final Field oneToManyAssociationFiled = entityMeta.getOneToManyAssociationFiled();
+//        Enhancer enhancer = new Enhancer();
+//        enhancer.setSuperclass(oneToManyAssociationFiled.getType());
+//        enhancer.setCallback((LazyLoader)() -> callback(););
+//
+//        return instance;
+//    }
+
+
+}
