+++ src/main/java/persistence/entity/loader/SimpleEntityLoader.java
@@ -1,0 +1,43 @@
+package persistence.entity.loader;
+
+import java.util.List;
+import jdbc.JdbcTemplate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import persistence.entity.EntityLoader;
+import persistence.sql.QueryGenerator;
+
+public class SimpleEntityLoader implements EntityLoader {
+    private final Logger log = LoggerFactory.getLogger(SimpleEntityLoader.class);
+    private final JdbcTemplate jdbcTemplate;
+    private final QueryGenerator queryGenerator;
+    private final EntityMapper entityMapper;
+
+    public SimpleEntityLoader(JdbcTemplate jdbcTemplate, QueryGenerator queryGenerator, EntityMapper entityMapper) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.queryGenerator = queryGenerator;
+        this.entityMapper = entityMapper;
+    }
+
+    public <T> T find(Class<T> tClass, Object id) {
+        final String query = queryGenerator
+                .select()
+                .findByIdQuery(id);
+
+        log.info(query);
+
+        return jdbcTemplate.queryForObject(query,
+                (resultSet) -> entityMapper.resultSetToEntity(tClass, resultSet));
+    }
+
+    public <T> List<T> findAll(Class<T> tClass) {
+        final String query = queryGenerator.select().findAllQuery();
+
+        log.info(query);
+
+        return jdbcTemplate.query(query,
+                (resultSet) -> entityMapper.resultSetToEntity(tClass, resultSet));
+    }
+
+
+}
