+++ src/test/java/persistence/entity/OneToManyAssociationTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.meta.EntityMeta;
+import persistence.testFixtures.assosiate.LazyLoadOrder;
 import persistence.testFixtures.assosiate.NoHasJoinColumnOrder;
 import persistence.testFixtures.assosiate.Order;
 
@@ -22,6 +23,7 @@ class OneToManyAssociationTest {
 
         //then
         assertSoftly((it)-> {
+            assertThat(oneToManyColumn.isLazy()).isFalse();
             assertThat(oneToManyColumn.isHasJoinColumn()).isTrue();
             assertThat(manyEntityMeta).isNotNull();
             assertThat(manyEntityMeta.getTableName()).isEqualTo("order_items");
@@ -44,4 +46,24 @@ class OneToManyAssociationTest {
             assertThat(manyEntityMeta.getTableName()).isEqualTo("order_items");
         });
     }
+
+    @Test
+    @DisplayName("oneToMany fetch type이 lazy인 생성한다")
+    void oneToManyFetch() throws Exception {
+        //given
+        OneToManyAssociation oneToManyColumn = OneToManyAssociation.of(LazyLoadOrder.class, EntityMeta.from(LazyLoadOrder.class));
+
+        //when
+        final EntityMeta manyEntityMeta = oneToManyColumn.getManyEntityMeta();
+
+        //then
+        assertSoftly((it)-> {
+            assertThat(oneToManyColumn.isLazy()).isTrue();
+            assertThat(oneToManyColumn.isHasJoinColumn()).isTrue();
+            assertThat(manyEntityMeta).isNotNull();
+            assertThat(manyEntityMeta.getTableName()).isEqualTo("order_items");
+        });
+    }
+
+
 }
