+++ src/test/java/persistence/testFixtures/assosiate/LazyLoadOrder.java
@@ -1,14 +1,19 @@
 package persistence.testFixtures.assosiate;
 
+
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
+import java.util.List;
 
 @Entity
 @Table(name = "orders")
-public class NoOneToManyOrder {
+public class LazyLoadOrder {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -16,5 +21,7 @@ public class NoOneToManyOrder {
 
     private String orderNumber;
 
-
+    @OneToMany(fetch = FetchType.LAZY)
+    @JoinColumn(name = "order_id")
+    private List<OrderItem> orderItems;
 }
